made the moabjobcontroller backwards compadible.
The last commit required that the os and machine type be in the default
yaml. This breaks old setups with the new version. I have now tried to
implement safe options for moab (like slurm) so that if an option does
not exist but is not nessary it still runs.