Parallelize test filter selection.
Encapsulate the main loop over paths in dir_db.select_from to transform and filter the paths in parallel.

Two consequences of this change motivated the other changes.
First, functions declared within other functions can't be pickled.  So move filters.filter_test_run out of the make_test_run_filter function and return it with the same arguments using functools.partial and the make_test_run_filter arguments.
Second, lambdas can't be pickled. So let order_func be a function defined in the top level of the filters module, sort_func, rather than a lambda stored as a value in a dictionary. This eliminates the SERIES_SORT_FUNC and TEST_SORT_FUNC dictionaries and replaces them with a dictionary with keys TEST and SERIES and values that are valid sort keys for these types of dictionaries. The order_func acts the same as previously, but any function that took the SORT_FUNC dicts as inputs now take a string (TEST or SERIES) instead.