Parallelize results query. (#420)
* Parallelize results query.

Add file result_utils.py to provide an analogue of status_utils.py for the results command.
The run method from the result.py plugin now calls get_results rather than initializing TestRun.load itself.  Also, get_results returns the test.results dict rather than a TestRun object (which can't be returned by a parallel map invocation because in that case pickle and the TestRun object don't get along.  A result_log field is added to that dict in case show_log is True.  This also simplifies the run command.  The update_results method is not parallel and still requires a full TestRun object so the argument tests is changed to test_ids, a list of ints, which are then initialized in update_results.

* Remove psutil and unnecessary imports.  Add conservative number of cpus estimate to config.py.

* Modified to conform to style conventions.

* Update result_utils.py

Co-authored-by: Paul Ferrell <51765748+Paul-Ferrell@users.noreply.github.com>