Added mutliprocessing to status query. (#418)
* Added mutliprocessing to status query.

In status_utils.py encapsulate the main loop in a separate function, get_status, to make it easier to use with map.  Use functools partial to reduce get_status to a single input to make map easier (could use starmap, but that seems more complicated). TestRun can't be picked in python < 3.7 because of some threading capabilities in it, so don't run in parallel unless python is new enough to work. Use psutils to get cpu_count to avoid using virtual cores and oversaturating the (front-end) node.

Improve .gitignore to allow users to store temporary and info files locally in .local, and avoid pushing compiled cache files.

* Remove psutil.  Add conservative number of cpus estimate to config.py and use to set processes in status_utils.

* Modified to conform to style conventions.

Co-authored-by: Paul Ferrell <51765748+Paul-Ferrell@users.noreply.github.com>