Partially fix Opt and Sopt plugins
Opt and Sopt plugins are necessary for generating command line
options generated from mode config files. Both Opt and Sopt take
an arguments which can be either a string or a list of strings,
which complicates their arg specs. This commit:

- Adds test cases for both Opt and Sopt plugins.
- Introduces a UnionSpec class to support arguments that are
  a union of multiple types.
- Begins to integrate UnionSpec into argument validation code.

Note that argument validation has not been fully implemented for
UnionSpec, which means that the added tests are not yet passing.