Finish fixing Opt and Sopt function plugins
Opt and Sopt plugins now have correct behavior for lists of strings
and pass all associated tests. Because both plugins can take either
a single string or a list of strings, this commit also finishes
implementing the UnionSpec specification (introduced in the previous
commit), which allows for function plugins to have parameters whose
actual arguments can take on one of several types. This specification
object is designed in the same spirit of the recently introduced Opt
specification, and will hopefully be useful for future plugins.

This commit also modifies the argument validation logic for function
plugins to account for the UnionSpec specification and adds unit tests
as well.