Modify transfomer to return function over a dict
This commit modifies the parse tree transformer to return an
anonymous function that takes in an arbitrary dictionary object
and returns a boolean. I don't know how to obtain a list of all
possible keys for the input dictionary, so at present, the
transformer simply tries to get the value associated with the
key (and returns None if the key is not present). Obviously,
in the code's present state, this will cause runtime errors. Future
commits will address this issue.